IcingaDB: Fix enumeral mismatch compiler warnings
This fixes 3 enumeral mismatch compiler warnings when building the
IcingaDB target.

+++ lib/icingadb/icingadb-objects.cpp
@@ -1219,7 +1219,7 @@ void IcingaDB::SendStatusUpdate(const ConfigObject::Ptr& object, const CheckResu
 	if (!cr) {
 		hard_state = 99;
 	} else {
-		hard_state = service ? service->GetLastHardState() : host->GetLastHardState();
+		hard_state = service ? static_cast<unsigned int>(service->GetLastHardState()) : static_cast<unsigned int>(host->GetLastHardState());
 	}
 
 	std::vector<String> xAdd ({
@@ -1228,7 +1228,7 @@ void IcingaDB::SendStatusUpdate(const ConfigObject::Ptr& object, const CheckResu
 		"environment_id", SHA1(GetEnvironment()),
 		"host_id", GetObjectIdentifier(host),
 		"state_type", Convert::ToString(type),
-		"soft_state", Convert::ToString(cr ? service ? cr->GetState() : Host::CalculateState(cr->GetState()) : 99),
+		"soft_state", Convert::ToString(cr ? service ? static_cast<unsigned int>(cr->GetState()) : static_cast<unsigned int>(Host::CalculateState(cr->GetState())) : 99),
 		"hard_state", Convert::ToString(hard_state),
 		"attempt", Convert::ToString(checkable->GetCheckAttempt()),
 		"previous_soft_state", Convert::ToString(GetPreviousState(checkable, service, StateTypeSoft)),
@@ -1304,7 +1304,7 @@ void IcingaDB::SendSentNotification(
 		"notification_id", GetObjectIdentifier(notification),
 		"host_id", GetObjectIdentifier(host),
 		"type", Convert::ToString(type),
-		"state", Convert::ToString(service ? cr->GetState() : Host::CalculateState(cr->GetState())),
+		"state", Convert::ToString(service ? static_cast<unsigned int>(cr->GetState()) : static_cast<unsigned int>(Host::CalculateState(cr->GetState()))),
 		"previous_hard_state", Convert::ToString(GetPreviousState(checkable, service, StateTypeHard)),
 		"author", Utility::ValidateUTF8(author),
 		"text", Utility::ValidateUTF8(finalText),
